Breaking the FluCam screen controls into their own mixin.
Changing from "removed" to "retracted" to align with what it says on the microscope.
